Fixed an Error in the Seed Method and Applied Code Reformatting
There was an error in the Seed method that was keeping the application from building the database correctly. Upon further investigation, I was using Async Methods that were not in an async method. These methods needed to have the .Wait() method added to them so that other executions to the same DbContext were not being performed. I plan on going back and making this method async, but that is going to involve a lot of testing. Basically, the issue was due to thread locks on the database when the next user was being seeded to the database.